Fix robust inputAge (with ignoring)
Just checking the bool conversion of "in" in the while condition does not work. The state of the stream is cleared at the end of the loop, i.e. no flag set. in.ignore does not necessarily set the fail bit when the stream is EOF. good() checks EOF though, while the bool conversion does not.

Condition that applies: end of file conditions occurs in the input sequence, in which case the function calls setstate(eofbit)